Merge separate SDL & VPI functionality into working simulator.
In kinda hacky and ugly way.

This makes me feel un-warm and un-fuzzy inside.

I would like to abstract all the board-specific stuff into a separate
interface that the other code introspects (to be able to easily support
multiple board types in the future) and cleanly separate the SDL from
VPI stuff.